refactor: Put redux files in redux folder Ticket: TISNEW-3985
