+++ src/redux/containers/Profile.tsx
@@ -1,9 +1,8 @@
 import React from "react";
 import { connect } from "react-redux";
-import ProgrammesComponent from "../programmes/Programmes";
-import PersonalDetailsComponent from "../personalDetails/PersonalDetails";
+import PersonalDetailsComponent from "../../components/personalDetails/PersonalDetails";
 import styles from "./Profile.module.scss";
-import { fetchPersonDetails } from "../../actions/personActions";
+import { fetchPersonDetails } from "../actions/personActions";
 
 class ProfileComponent extends React.PureComponent<any> {
   componentDidMount() {
@@ -12,18 +11,22 @@ class ProfileComponent extends React.PureComponent<any> {
 
   render() {
     const personDetails = this.props.personReducer.personalDetails;
-    if (personDetails !== undefined) {
+    const isLoaded = this.props.personReducer.isLoaded;
+    const error = this.props.personReducer.error;
+
+    if (error) {
+      return <div>Error: {error}</div>;
+    } else if (!isLoaded) {
+      return <div>Loading...</div>;
+    } else {
       return (
         personDetails && (
           <div className={styles.profileContainer}>
             <PersonalDetailsComponent personalDetail={personDetails} />
-            <ProgrammesComponent
-              programmeMemberships={personDetails.programmeMemberships}
-            />
           </div>
         )
       );
-    } else return <div>Nothing coming through</div>;
+    }
   }
 }
 
@@ -31,7 +34,9 @@ const mapStateToProps = (state: {
   personReducer: { personalDetails: any; isLoaded: boolean; error: any };
 }) => ({
   personReducer: {
-    personalDetails: state.personReducer.personalDetails.data
+    personalDetails: state.personReducer.personalDetails.data,
+    isLoaded: state.personReducer.isLoaded,
+    error: state.personReducer.error
   }
 });
